backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-allauth>=0.57
django-cors-headers>=4.3
django-csp>=3.7
django-filter>=23.3
django-storages>=1.14
django-redis>=5.4
django-encrypted-model-fields>=0.6
django-celery-beat>=2.5
psycopg2-binary>=2.9
redis>=5.0
celery>=5.3
stripe>=7.0
boto3>=1.29
pyotp>=2.9
qrcode>=7.4
Pillow>=10.1
cryptography>=41.0
gunicorn>=21.2
whitenoise>=6.6
python-decouple>=3.8
requests>=2.31
django-ratelimit>=4.1
